Updating check for child / parent user, text
